+++ libraries/base-testing-tool/lib/TestingToolModule.ts
@@ -16,15 +16,15 @@
  * limitations under the License.
  */
 import { Module, ModuleManager } from "@syntest/module";
-import { SimpleObjectiveManagerPlugin } from "./plugins/objectiveManagers/SimpleObjectiveManagerPlugin";
-import { StructuralObjectiveManagerPlugin } from "./plugins/objectiveManagers/StructuralObjectiveManagerPlugin";
-import { UncoveredObjectiveManagerPlugin } from "./plugins/objectiveManagers/UncoveredObjectiveManagerPlugin";
+import { SimpleObjectiveManagerPlugin } from "./plugins/objective-managers/SimpleObjectiveManagerPlugin";
+import { StructuralObjectiveManagerPlugin } from "./plugins/objective-managers/StructuralObjectiveManagerPlugin";
+import { UncoveredObjectiveManagerPlugin } from "./plugins/objective-managers/UncoveredObjectiveManagerPlugin";
 import { DefaultProcreationPlugin } from "./plugins/procreation-operators/DefaultProcreationPlugin";
-import { MOSAFamilyPlugin } from "./plugins/searchAlgorithms/MOSAFamilyPlugin";
-import { NSGAIIPlugin } from "./plugins/searchAlgorithms/NSGAIIPlugin";
-import { RandomSearchPlugin } from "./plugins/searchAlgorithms/RandomSearchPlugin";
-import { SignalTerminationTriggerPlugin } from "./plugins/terminationTriggers/SignalTerminationTriggerPlugin";
-import { LengthObjectiveComparatorPlugin } from "./plugins/secondaryObjectives/LengthObjectiveComparatorPlugin";
+import { MOSAFamilyPlugin } from "./plugins/search-algorithms/MOSAFamilyPlugin";
+import { NSGAIIPlugin } from "./plugins/search-algorithms/NSGAIIPlugin";
+import { RandomSearchPlugin } from "./plugins/search-algorithms/RandomSearchPlugin";
+import { SignalTerminationTriggerPlugin } from "./plugins/termination-triggers/SignalTerminationTriggerPlugin";
+import { LengthObjectiveComparatorPlugin } from "./plugins/secondary-objectives/LengthObjectiveComparatorPlugin";
 import { SearchMetricListener } from "./plugins/listeners/SearchMetricListener";
 import { MOSAPreset } from "./presets/MOSAPreset";
 import { DynaMOSAPreset } from "./presets/DynaMOSAPreset";
