+++ libraries/base-testing-tool/lib/plugins/objective-managers/SimpleObjectiveManagerPlugin.ts
