+++ libraries/base-testing-tool/lib/plugins/objective-managers/StructuralObjectiveManagerPlugin.ts
