+++ libraries/base-testing-tool/lib/plugins/objective-managers/UncoveredObjectiveManagerPlugin.ts
