+++ libraries/base-testing-tool/lib/plugins/termination-triggers/SignalTerminationTriggerPlugin.ts
